pyproject.toml
[project]
name = "nonebot-plugin-tg-stickers-downloads"
version = "0.1.0"
description = "Telegram 贴纸包下载插件"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "pydantic",
    "aiofiles>=24.1.0,<26.0.0",
    "anyio>=4.0.0,<5.0.0",
    "httpx>=0.28.0,<1.0.0",
    "nonebot2>=2.2.1,<3.0.0",
    "nonebot-plugin-alconna>=0.57.1,<1.0.0",
    "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
    "nonebot-plugin-localstore>=0.7.0,<1.0.0",
    "pillow>=10.0.0,<13.0.0",
    "rlottie_python>=1.3.8,<2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
    "ty>=0.0.26",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["src/dragon_contest/migrations/**"]

[tool.ruff.lint]
select = [
    "I",
    "E", "W", "F",
    "UP",
    "B",
    "C4", "C90",
    "SIM",
    "PERF",
    "RUF", "RUF100",
    "PT",
    "Q",
]

ignore = ["I001","E402", "B008", "RUF001"]
fixable = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
